requirements.txt
amqp==2.6.1
asgiref==3.5.2
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
colorclass==2.2.2
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.16
django-allauth==0.51.0
django-autoslug-iplweb==1.9.5
django-categories==1.9.2
django-ckeditor==6.0.0
django-classy-tags==3.0.1
django-cms==3.7.4
django-filter==22.1
django-formtools==2.4
django-google-analytics-app==5.0.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-nine==0.2.5
django-object-tools==2.0.0
django-sekizai==2.0.0
django-seo==0.3.5
django-smartfields==1.1.3
django-tinymce==3.0.2
djangocms-admin-style==3.2.0
djangorestframework==3.14.0
docopt==0.6.2
easy-thumbnails==2.8.3
feedparser==6.0.10
gunicorn==20.1.0
idna==3.4
importlib-metadata==3.3.0
kombu==4.6.11
latex2mathml==3.63.2
Markdown==3.3.4
mpmath==1.2.1
mysqlclient==2.0.3
npm==0.1.1
oauthlib==3.2.2
optional-django==0.3.0
ordereddict==1.1
packaging==21.3
pif==0.8.2
Pillow==9.3.0
pip-upgrader==1.4.15
psycopg2==2.8.6
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytexit==0.3.4
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.1
sqlparse==0.4.3
structlog==20.2.0
sweetify==2.3.1
sympy==1.9
terminaltables==3.1.10
transliterate==1.10.2
typing-extensions==3.7.4.3
unicode-slugify==0.1.5
Unidecode==1.3.6
urllib3==1.26.12
vine==1.3.0
vishap==0.1.5
whitenoise==6.2.0
zipp==3.4.0
